# Use char instead of byte indexes for length, substr, index, and match