--- a/goawk.go
+++ b/goawk.go
@@ -44,7 +44,7 @@ import (
 )
 
 const (
-	version    = "v1.9.2"
+	version    = "v1.11.0"
 	copyright  = "GoAWK " + version + " - Copyright (c) 2021 Ben Hoyt"
 	shortUsage = "usage: goawk [-F fs] [-v var=value] [-f progfile | 'prog'] [file ...]"
 	longUsage  = `Standard AWK arguments:
@@ -56,6 +56,7 @@ const (
         load AWK source from progfile (multiple allowed)
 
 Additional GoAWK arguments:
+  -b    use byte indexes for index(), length(), match(), and substr()
   -cpuprofile file
         write CPU profile to file
   -d    debug mode (print parsed AST to stderr)
@@ -77,6 +78,7 @@ func main() {
 	debug := false
 	debugTypes := false
 	memprofile := ""
+	useBytes := false
 
 	var i int
 	for i = 1; i < len(os.Args); i++ {
@@ -109,6 +111,8 @@ func main() {
 			}
 			i++
 			vars = append(vars, os.Args[i])
+		case "-b":
+			useBytes = true
 		case "-cpuprofile":
 			if i+1 >= len(os.Args) {
 				errorExitf("flag needs an argument: -cpuprofile")
@@ -212,6 +216,7 @@ func main() {
 	config := &interp.Config{
 		Argv0: filepath.Base(os.Args[0]),
 		Args:  expandWildcardsOnWindows(args),
+		Bytes: useBytes,
 		Vars:  []string{"FS", fieldSep},
 	}
 	for _, v := range vars {