--- a/goawk_test.go
+++ b/goawk_test.go
@@ -353,7 +353,7 @@ func TestCommandLine(t *testing.T) {
 
 		// Byte index vs character index mode
 		{[]string{`{ print length }`}, "絵\n", "1\n", ""},
-		{[]string{"-b", `{ print length }`}, "絵\n", "3\n", ""},
+		{[]string{"-b", `{ print length }  # !windows-gawk`}, "絵\n", "3\n", ""},
 
 		// ARGV/ARGC handling
 		{[]string{`
@@ -444,6 +444,11 @@ func runGoAWK(args []string, stdin string) (stdout, stderr string, err error) {
 }
 
 func runAWKs(t *testing.T, testArgs []string, testStdin, testOutput, testError string) {
+	for _, arg := range testArgs {
+		if strings.Contains(arg, "!"+runtime.GOOS+"-"+awkExe) {
+			t.Skipf("skipping on %s under %s", runtime.GOOS, awkExe)
+		}
+	}
 	cmd := exec.Command(awkExe, testArgs...)
 	if testStdin != "" {
 		cmd.Stdin = bytes.NewReader([]byte(testStdin))