@benhoyt 2022-01-03
FWIW, the "real world" benchmark comparison is as follows (`orig` is the master branch, v1.10.0, and `goawk` is this branch):

```
Test      | goawk |  orig |   awk 
--------- | ----- | ----- | -----
tt.01     |  3.53 |  2.65 |  1.00
tt.02     |  1.85 |  1.82 |  1.00
tt.02a    |  1.81 |  1.90 |  1.00
tt.03     |  1.16 |  1.36 |  1.00
tt.03a    |  1.42 |  1.34 |  1.00
tt.04     |  0.84 |  0.85 |  1.00
tt.05     |  0.98 |  0.96 |  1.00
tt.06     |  1.27 |  1.44 |  1.00
tt.07     |  1.46 |  1.45 |  1.00
tt.08     |  1.85 |  2.59 |  1.00
tt.09     |  2.33 |  2.29 |  1.00
tt.10     |  0.39 |  0.40 |  1.00
tt.10a    |  0.37 |  0.36 |  1.00
tt.11     |  2.15 |  2.38 |  1.00
tt.12     |  1.41 |  1.42 |  1.00
tt.13     |  2.69 |  2.63 |  1.00
tt.13a    |  1.92 |  1.92 |  1.00
tt.14     |  1.18 |  1.18 |  1.00
tt.15     |  0.61 |  0.65 |  1.00
tt.16     |  1.27 |  1.30 |  1.00
tt.big    |  1.62 |  1.61 |  1.00
tt.x1     |  1.53 |  1.56 |  1.00
tt.x2     |  0.99 |  0.98 |  1.00
--------- | ----- | ----- | -----
**Geo mean**  | **1.43** | **1.43** | **1.00**
```

So overall I'm pretty happy with how this turned out. Thanks @arnoldrobbins for pushing on this. :-)